Change i2c speed to 400k for this board
It looks like 100kHz is problematic in the demo board as without some
small (~5ms) delays the read operation fails with an ARLO error.
Bumping to 400kHz or adding those small delays works.
The problem is specailly notorius when using the bareboard and a release
binary. It seems that a small delay, like unoptimized code makes i2c
work in the board at 100kHz. This would require some bus smelling to RCA
the problem properly.